Add useMinimalVideoPlayback hook and cleanup
