+++ .changeset/calm-videos-play.md
@@ -2,7 +2,7 @@
 '@primer/react-brand': minor
 ---
 
-Added the new `MinimalVideoPlayer` component for presenting muted, looping videos with an accessible play and pause control. It automatically pauses when outside the viewport and respects reduced motion preferences.
+Added the new `MinimalVideoPlayer` component for presenting muted, looping videos with a minimal play and pause control.
 
 ```tsx
 <MinimalVideoPlayer poster="/video-poster.jpg" src="/video.webm" title="Product interface demonstration" />
